requirements.txt
flask
keras
numpy
pandas
scikit-learn
tensorflow
gunicorn
flask-cors
python-dotenv